+++ modeltrans/management/commands/i18n_make_indexes.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import print_function, unicode_literals
 
-from collections import defaultdict
-
 from django.core.management.base import BaseCommand
 
 from modeltrans.translator import get_translated_models