+++ modeltrans/utils.py
@@ -1,8 +1,6 @@
 import hashlib
 
-from django.utils import six
-from django.utils.encoding import force_bytes, force_text
-from django.utils.functional import lazy
+from django.utils.encoding import force_bytes
 from django.utils.translation import get_language as _get_language
 
 from .conf import get_available_languages, get_default_language