+++ CHANGELOG.md
@@ -5,6 +5,8 @@
  - Any translated fields used in `Model.Meta.ordering` is transformed into the correct expression with django 2.0 and later (fixes #25).
  - `django.contrib.postgres.GinIndex` is added to the `i18n` column if it's supported by the django version used (1.11 and later). It can be disabled with the setting `MODELTRANS_CREATE_GIN`.
  - The migration generated from `./manage.py i18n_makemigrations <app>` used to move the data and add a GIN index. This is split into two commands: `./manage.py i18n_makemigrations` and `./manage.py i18n_make_indexes`.
+ - Added support for `values(**expressions)`` with references to translated fields.
+ - Added support for translated values in `annotate()`
 
 ## 0.1.2 (2017-10-23)
  - Ensure a dynamic mixed `MultilingualQuerySet` can be pickled.